ignore package lock
my reasoning may be flawed here, but by remove the package lock, I _think_ this means that users will always get the latest version all-the-public-replicate-models, as that's specified as `*` in package.json. The API for that doesn't change. It just accumulates more models each day as they're published to Replicate.